refactor(appkit): extract agents skill loading/dispatch into skill-loader
Step 2 of splitting the agents plugin. Moves skill discovery, per-agent
catalog resolution, and the load_skill/read_skill_file dispatch into
skill-loader.ts as free functions; the class keeps thin delegators (call sites
unchanged) and skillWorkspaceClient() as the OBO credential seam.

agents.ts: 2331 -> 2152 lines. typecheck + 394 agent tests green.
